extraction complete, notification class removed
