+++ Turfy/Turfy/AppDelegate.swift
@@ -11,15 +11,16 @@ import CoreLocation
 import Firebase
 import FBSDKCoreKit
 import FBSDKLoginKit
+import UIKit
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
 
 	
 	var window: UIWindow?
-	var messages: [Message] = []
     let locationManager = CLLocationManager()
-	
+    var messages : [Message] = []
+    
 	func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
 		
         // Override point for customization after application launch.
@@ -96,6 +97,8 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         return "I AM AWSOME!!!"
         
     }
+    
+ 
 
 }
